% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_counts}
\alias{avg_counts}
\title{Example dataset with average expression of nine tissues in five
different species}
\format{
An object of class \code{'cross'}; see \code{\link[qtl]{read.cross}}.
}
\usage{
avg_counts <- SBF::avg_counts
}
\description{
Example dataset with average expression of nine tissues in five
different species
}
\examples{
avg_counts <- SBF::avg_counts
# species names
names(avg_counts)
}
\keyword{datasets}
