% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calcAvgCounts}
\alias{calcAvgCounts}
\title{Function to calculate mean expression values for each tissue/group}
\usage{
calcAvgCounts(counts, metadata, ndecimal = 4)
}
\arguments{
\item{counts}{gene expression matrix}

\item{metadata}{data frame with species tissue key column}

\item{ndecimal}{number of decimals to round}
}
\value{
a list with mean expression
}
\description{
Function to calculate mean expression values for each tissue/group
}
