% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{calcNearestOrthoMatrix}
\alias{calcNearestOrthoMatrix}
\title{Find Nearest orthogonal matrix}
\usage{
calcNearestOrthoMatrix(mat_list)
}
\arguments{
\item{mat_list}{A list of numeric matrices}
}
\value{
orthogonal matrix
}
\description{
Function to compute nearest orthogonal matrix to a set of matrices.
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 3, ncols = 3, nrows = 3)
nearest_ortho <- calcNearestOrthoMatrix(mymat)
print(nearest_ortho \%*\% t(nearest_ortho))
}
