% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calcPercentInfo}
\alias{calcPercentInfo}
\title{Compute information represented by each dimensions}
\usage{
calcPercentInfo(l)
}
\arguments{
\item{l}{SBF list object}
}
\value{
list with percentage of information represented by each dimension
according to their Delta_i values
}
\description{
Compute information represented by each dimensions
}
\examples{
#' @examples
# create test dataset
set.seed(1231)
mymat <- createRandomMatrices(n = 4, ncols = 3, nrows = 4:6)

# SBF call. Estimate V using the sum of Di^TDi
sbf <- SBF(matrix_list = mymat)
calcPercentInfo(sbf)
}
