% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{computeUDelta}
\alias{computeUDelta}
\title{Compute U and Delta for a given set of matrices and a V}
\usage{
computeUDelta(mat_list, V, approximate = TRUE)
}
\arguments{
\item{mat_list}{A list containing numeric matrices}

\item{V}{A numeric matrix of dimension ncol(matrix1) * ncol(matrix1)}

\item{approximate}{Compute U with orthonormal columns. Default TRUE}
}
\value{
list of U and delta
}
\description{
Function to compute U and Delta for a given set of matrices and V.
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 3, ncols = 3, nrows = 3)
set.seed(325)
mymat_rand <- createRandomMatrices(n = 1, ncols = 3, nrows = 3)[[1]]
out <- computeUDelta(mymat, svd(mymat_rand)$v)
}
