% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createRandomMatrices}
\alias{createRandomMatrices}
\title{Generate random integer matrices with full column rank}
\usage{
createRandomMatrices(n = 3, ncols = 3, nrows = 3:6, max_iter = 1000)
}
\arguments{
\item{n}{number of matrices to generate. Integer value. Default 3.}

\item{ncols}{number of columns for each matrix. All matrices will have
the same number of columns. Integer value. Default 3.}

\item{nrows}{possible values for the number of rows for the matrices.
The number of rows should be greater than the column value.
Default c(3,4,5,6).}

\item{max_iter}{maximum number of iteration. Default 1000}
}
\value{
a list containing random matrices
}
\description{
Function to generate random matrices with user-specified columns
and rows
}
\examples{
# generate 4 matrices with possible rows = c(4,8) with 3 columns
set.seed(123)
mymat <- createRandomMatrices(n = 4, ncols = 3, nrows = 4:8)
}
