% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getScientificName}
\alias{getScientificName}
\title{Function to return scientific name of species}
\usage{
getScientificName(x)
}
\arguments{
\item{x}{species name}
}
\value{
species name in scientific format
}
\description{
Function to return scientific name of species
}
\examples{
species <- "Homo_sapiens"
getScientificName(species)
}
