% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getSpeciesShortName}
\alias{getSpeciesShortName}
\title{Function to create short name for species}
\usage{
getSpeciesShortName(species_list)
}
\arguments{
\item{species_list}{a vector of species name}
}
\value{
short version of species name
}
\description{
Function to create short name for species
}
\examples{
species <- c("Homo_sapiens", "Mus_musculus")
species_short <- sapply(species, getSpeciesShortName)
species <- "Homo_sapiens"
getSpeciesShortName(species)
}
