% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{GSVD}
\alias{GSVD}
\title{Compute GSVD factorization}
\usage{
GSVD(D1, D2)
}
\arguments{
\item{D1}{A numeric matrix for GSVD factorization.}

\item{D2}{A numeric matrix for GSVD factorization.}
}
\value{
a list containing u1, u2, v, d1, and d2
}
\description{
Function to compute GSVD factorization for two matrices D1 and D2
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 2, ncols = 3, nrows = 4:6)
gsvd <- GSVD(mymat$mat1, mymat$mat2)
print(gsvd$v)
}
