% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{HOGSVD}
\alias{HOGSVD}
\title{Compute HOGSVD factorization}
\usage{
HOGSVD(
  matrix_list = NULL,
  check_col_matching = FALSE,
  col_index = NULL,
  col_sep = "_",
  verbose = FALSE
)
}
\arguments{
\item{matrix_list}{A list containing Di matrices for joint matrix
factorization. Column names of each Di matrix may or may not have information
about tissue or cell type.}

\item{check_col_matching}{if the column names have information about tissue/
cell type and one-to-one correspondence of tissue types across species has to
be checked, set this parameter to be TRUE. Default FALSE.}

\item{col_index}{If a separator separates information in column names,
the col_index is the index in the column name corresponding to tissue or
cell type. E.g. for column name 'hsapiens_brain', col_index is 2.
Only checked if check_col_matching = TRUE. Default NULL.}

\item{col_sep}{separator in column names to separate different fields.
E.g. for column names 'hsapiens_brain', 'hsapiens_heart' etc., the separator
is underscore. Set it to NULL if column matching across species has to be
performed and there is no separator in the column names.
Only checked if check_col_matching = TRUE. Default underscore.}

\item{verbose}{if TRUE print verbose lines. Default FALSE.}
}
\value{
a list containing u, v, lambda, d, and other outputs
of HOGSVD factorization.
}
\description{
Function to compute HOGSVD factorization
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 4, ncols = 3, nrows = 4:6)
hogsvd <- HOGSVD(matrix_list = mymat)
print(hogsvd$v)
}
