% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{invertDelta}
\alias{invertDelta}
\title{Invert delta_i entries and return a matrix}
\usage{
invertDelta(l)
}
\arguments{
\item{l}{SBF list object}
}
\value{
a matrix with inverted delta_i values
}
\description{
Invert delta_i entries and return a matrix
}
