% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{normalizeTPM}
\alias{normalizeTPM}
\title{TPM normalization for a gene expression matrix}
\usage{
normalizeTPM(rawCounts, gene_len = NULL)
}
\arguments{
\item{rawCounts}{Raw gene expression counts matrix}

\item{gene_len}{Matching gene lengths (kb)}
}
\value{
TPM normalized matrix
}
\description{
TPM normalization for a gene expression matrix
}
