% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{projectCounts}
\alias{projectCounts}
\title{Function to project counts to the common expression space}
\usage{
projectCounts(counts, obj, combine = TRUE)
}
\arguments{
\item{counts}{mean expression list or counts list}

\item{obj}{SBF output list}

\item{combine}{If multiple libraries exist, the function
will row bind the projected counts if set TRUE}
}
\value{
projected counts data.frame if combine is set TRUE else
list with projected counts
}
\description{
Function to project counts to the common expression space
}
