\name{sBF-package}
\Rdversion{1.1}
\alias{sBF-package}
\docType{package}
\title{
Smooth Backfitting Estimator Package
}
\description{
Smooth Backfitting Estimator
}
\details{
\tabular{ll}{
Package: \tab sBF\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-24\cr
License: \tab GPL (>= 2) \cr
}
}
\author{
A. Arcagni \url{a.arcagni@campus.unimib.it} and L. Bagnato \url{luca.bagnato@unimib.it}
}
\references{
T. Hastie and R. J. Tibshirani. \emph{Generalised additive models}. Chapman and Hall: London, 1990.

E. Mammen, O. Linton, and J. Nielsen. \emph{The existence and asymptotic properties of a backfitting projection algorithm under weak conditions}. The Annals of Statistics, 27(5):1443-1490, 1999.

J. P. Nielsen and S. Sperlich. \emph{Smooth backfitting in practice}. Journal of the Royal Statistical Society, Series B: Statistical Methodology, 67(1):43-61, 2005.
}
\seealso{
\code{\link[sBF]{sBF}}, \code{\link[sBF]{K}}.
}
