% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TissueExprSpecies}
\alias{TissueExprSpecies}
\title{Gene expression dataset of five tissues in three
different species}
\format{
A list containing gene expression data frame from three species.
}
\usage{
avg_counts <- SBF::TissueExprSpecies
}
\description{
An example dataset containing average gene expression values of
five tissues in three different species.
}
\examples{
# load dataset
avg_counts <- SBF::TissueExprSpecies

# species names
names(avg_counts)

# dimension of matrices for different species
sapply(avg_counts, dim)

# names of different tissue types
names(avg_counts[[names(avg_counts)[1]]])
}
\keyword{datasets}
