% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{updateDelta}
\alias{updateDelta}
\title{Update Delta}
\usage{
updateDelta(mat_list, u, v)
}
\arguments{
\item{mat_list}{A list containing numeric matrices}

\item{u}{A list containing U matrices}

\item{v}{V matrix}
}
\value{
a list containing new U matrices
}
\description{
Function to update Delta using given Uis and V. The new Delta is set as
diag(U^T D V).
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 3, ncols = 3, nrows = 3)
sbf <- SBF(matrix_list = mymat, orthogonal = FALSE,
           transform_matrix = FALSE)
newU <- updateU(mymat, sbf$delta, sbf$v)
newDelta <- updateDelta(mymat, newU, sbf$v)
}
