% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{updateU}
\alias{updateU}
\title{Update U with orthonormal columns}
\usage{
updateU(mat_list, d, v)
}
\arguments{
\item{mat_list}{A list containing numeric matrices}

\item{d}{A list containing delta matrices}

\item{v}{V matrix}
}
\value{
a list containing new U matrices
}
\description{
Function to update U using given delta and V. The new U is set as XY^T,
where SVD of D V Delta = X Sigma Y^T.
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 3, ncols = 3, nrows = 3)
sbf <- SBF(matrix_list = mymat, orthogonal = FALSE,
           transform_matrix = FALSE)
newU <- updateU(mymat, sbf$delta, sbf$v)
}
