% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{updateV}
\alias{updateV}
\title{Update V}
\usage{
updateV(mat_list, u, d)
}
\arguments{
\item{mat_list}{A list containing numeric matrices}

\item{u}{A list containing U matrices}

\item{d}{A list containing delta matrices}
}
\value{
a list containing new U matrices
}
\description{
Function to update V using give Delta and U. The new V is set as
X Y^T, where SVD of D^T U Delta = X Sigma Y^T.
}
\examples{
set.seed(1231)
mymat <- createRandomMatrices(n = 3, ncols = 3, nrows = 3)
sbf <- SBF(matrix_list = mymat, approximate = FALSE,
           transform_matrix = FALSE)
newU <- updateU(mymat, sbf$delta, sbf$v)
newDelta <- updateDelta(mymat, newU, sbf$v)
newV <- updateV(mymat, newU, newDelta)
}
