% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_counts}
\alias{avg_counts}
\title{Datasets with average expression of 9 tissue types in 5 different species}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(avg_counts)
}
\description{
Datasets with average expression of 9 tissue types in 5 different species
}
\examples{
data(avg_counts)
# species names
names(avg_counts)
}
\keyword{datasets}
