% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcDecompError}
\alias{calcDecompError}
\title{Calculate decomposition error}
\usage{
calcDecompError(matrix_initial, delta_i, u_i, v_i)
}
\arguments{
\item{matrix_initial}{list with initial Di matrices}

\item{delta_i}{delta values computed using SBF/A-SBF function}

\item{u_i}{Ui values computed using SBF/A-SBF function}

\item{v_i}{V computed using SBF/A-SBF function}
}
\value{
a numeric value for the factorization error
}
\description{
Function to compute Forbenius norm between two sets of matrix
}
\examples{
# load dataset
avg_counts <- SBF::avg_counts
# call sbf
sbf <- SBF(matrix_list = avg_counts, col_index = 2, approximate = FALSE,
           transform_matrix = FALSE, verbose = FALSE)

# calculate decomposition error
decomperror <- calcDecompError(avg_counts, sbf$delta, sbf$u, sbf$v)

# e.g. 2
asbf_cor <- SBF(matrix_list = avg_counts, col_index = 2, approximate = FALSE,
               transform_matrix = TRUE, verbose = FALSE)
decomperror <- calcDecompError(avg_counts, asbf_cor$delta, asbf_cor$u_ortho,
                               asbf_cor$v)
}
