% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcDecompError}
\alias{calcDecompError}
\title{Calculate decomposition error}
\usage{
calcDecompError(matrix_initial, delta, u, v)
}
\arguments{
\item{matrix_initial}{list with initial Di matrices}

\item{delta}{delta_i values computed using SBF/A-SBF function}

\item{u}{U_i values computed using SBF/A-SBF function}

\item{v}{V computed using SBF/A-SBF function}
}
\value{
a numeric value for the factorization error
}
\description{
Function to compute Forbenius norm between two sets of matrix.
Computes sum of Forbenius norm for
\code{matrix_initial[[i]] - u[[i]]  diag(delta[[i]])  t(v)} for
all i.
}
\examples{
# load dataset
avg_counts <- SBF::TissueExprSpecies
# call sbf
sbf <- SBF(matrix_list = avg_counts, col_index = 2, approximate = FALSE,
           transform_matrix = FALSE, verbose = FALSE)

# calculate decomposition error
decomperror <- calcDecompError(avg_counts, sbf$delta, sbf$u, sbf$v)

# e.g. 2
avg_counts <- SBF::TissueExprSpecies
asbf_cor <- SBF(matrix_list = avg_counts, col_index = 2, approximate = TRUE,
               transform_matrix = TRUE, verbose = FALSE)
decomperror <- calcDecompError(avg_counts, asbf_cor$delta, asbf_cor$u_ortho,
                               asbf_cor$v)
}
