% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcDecompError}
\alias{calcDecompError}
\title{Calculate decomposition error (Forbenius norm)}
\usage{
calcDecompError(matrix.list.start, sig_i, u_i, v_i)
}
\arguments{
\item{matrix.list.start}{list with Di matrix}

\item{sig_i}{sigma values computed using SBF/A-SBF function}

\item{u_i}{Ui values computed using SBF/A-SBF function}

\item{v_i}{V computed using SBF/A-SBF function}
}
\value{

}
\description{
Calculate decomposition error (Forbenius norm)
}
\examples{
# load dataset
avg_counts <- SBF::avg_counts
# call sbf
sbf <- SBF(avg_counts = avg_counts, colIndex = 2, approximate = FALSE,
           transformData = FALSE, verbose = FALSE)

# calculate decomposition error
calcDecompError(avg_counts,sbf$sigma,sbf$u,sbf$v)$error

# e.g. 2
sbf.cor <- SBF(avg_counts = avg_counts, colIndex = 2, approximate = FALSE,
               transformData = TRUE, verbose = FALSE)
calcDecompError(avg_counts,sbf.cor$sigma,sbf.cor$u,sbf.cor$v)
}
