% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{countToTpm}
\alias{countToTpm}
\title{Calculate TPM normalization for a vector}
\usage{
countToTpm(x, gene_len)
}
\arguments{
\item{x}{a column of gene expression profiles}

\item{gene_len}{gene lengths (kb)}
}
\description{
Calculate TPM normalization for a vector
}
