% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF.R
\name{SBF}
\alias{SBF}
\title{Function to compute Shared Basis Factorization (SBF) and
Approximate Shared Basis Factorization (A-SBF)}
\usage{
SBF(
  avg_counts = NULL,
  check_col_matching = TRUE,
  colNameSep = "_",
  colIndex = NULL,
  approximate = TRUE,
  transformData = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{avg_counts}{A list containing Di matrices for joint factorization.
Column name of each Di matrix will have information about tissue/celltype
separated by delimitter:species_tissue/celltype_otherdetails. For e.g.
"hsapiens_brain"  "hsapiens_heart"  "hsapiens_kidney" "hsapiens_liver"
"hsapiens_testis".}

\item{check_col_matching}{Check whether tissues/cell types are matching in
different Di. Default TRUE}

\item{colNameSep}{separator in column names to separate different fields.
For e.g "hsapiens_brain"  "hsapiens_heart" separator is "\emph{". Default "}"}

\item{colIndex}{Index in column name corresponding to Tissue/Celltype.
For e.g. "hsapiens_brain" colIndex is 2. DEFAULT NULL}

\item{approximate}{TRUE will compute A-SBF. DEFAULT TRUE}

\item{transformData}{if TRUE then Di will be transfered to compute correlation
matrix and V is computed based on this. Otherwise V is computed using
Di^TDi. Default FALSE}

\item{verbose}{if TRUE print verbose lines. Default FALSE}
}
\value{

}
\description{
Function to compute Shared Basis Factorization (SBF) and
Approximate Shared Basis Factorization (A-SBF)
}
\examples{
data(avg_counts,package="SBF")
# SBF call
sbf = SBF(avg_counts = avg_counts,colIndex = 2, approximate = FALSE,
         transformData = FALSE)
# SBF call using correlation matrix
sbf.cor = SBF(avg_counts = avg_counts,colIndex = 2, approximate = FALSE,
transformData = TRUE)
# A-SBF call
asbf = SBF(avg_counts = avg_counts,colIndex = 2, approximate = TRUE,
transformData = FALSE)
# A-SBF call using correlation matrix
asbf.cor = SBF(avg_counts = avg_counts,colIndex = 2, approximate = TRUE,
transformData = TRUE)
}
