% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF.R
\name{SBF}
\alias{SBF}
\title{Function to compute Shared Basis Factorization (SBF) and
Approximate Shared Basis Factorization (A-SBF)}
\usage{
SBF(
  avg_counts = NULL,
  check_col_matching = TRUE,
  colNameSep = "_",
  colIndex = NULL,
  approximate = TRUE,
  transformData = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{avg_counts}{A list containing Di matrices for joint matrix
factorization. Column names of each Di matrix may or may not have information
about tissue or cell type.}

\item{check_col_matching}{if the column names have information about tissue/
cell type and one-to-one correspondence of tissue types across species has to
be checked, set this parameter to be TRUE. Otherwise, set it to FALSE.}

\item{colNameSep}{separator in column names to separate different fields.
E.g. for column names 'hsapiens_brain', 'hsapiens_heart' etc., the separator
is underscore. Set it to NULL if column matching across species has to be
performed and there is no separator in the column names.
Only checked if check_col_matching = TRUE. Default underscore.}

\item{colIndex}{If a separator separates information in column names,
the colIndex is the index in the column name corresponding to tissue or
cell type. E.g. for column name 'hsapiens_brain', colIndex is 2.
Only checked if check_col_matching = TRUE. Default NULL}

\item{approximate}{TRUE will compute A-SBF. Default TRUE}

\item{transformData}{if TRUE, then Di will be transformed to compute
correlation matrix, and V is computed based on this instead of
Di^TDi. An unbiased estimate of covariance (denominator n-1) is
used for the computing correlation. Default FALSE}

\item{verbose}{if TRUE print verbose lines. Default FALSE}
}
\value{

}
\description{
Function to compute Shared Basis Factorization (SBF) and
Approximate Shared Basis Factorization (A-SBF)
}
\examples{
# SBF call
avg_counts <- SBF::avg_counts
sbf <- SBF(avg_counts = avg_counts,colIndex = 2, approximate = FALSE,
         transformData = FALSE)

# SBF call using correlation matrix
avg_counts <- SBF::avg_counts
sbf.cor <- SBF(avg_counts = avg_counts,colIndex = 2, approximate = FALSE,
              transformData = TRUE)

# A-SBF call
avg_counts <- SBF::avg_counts
asbf <- SBF(avg_counts = avg_counts,colIndex = 2, approximate = TRUE,
           transformData = FALSE)

# A-SBF call using correlation matrix
avg_counts <- SBF::avg_counts
asbf.cor <- SBF(avg_counts = avg_counts,colIndex = 2, approximate = TRUE,
               transformData = TRUE)
}
