% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBF.R
\name{SBF}
\alias{SBF}
\title{Compute Shared Basis Factorization (SBF) and Approximate Shared Basis
Factorization (A-SBF)}
\usage{
SBF(
  matrix_list = NULL,
  check_col_matching = FALSE,
  col_sep = "_",
  col_index = NULL,
  weighted = FALSE,
  approximate = FALSE,
  transform_matrix = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{matrix_list}{A list containing Di matrices for joint matrix
factorization. Column names of each Di matrix may or may not have information
about tissue or cell type.}

\item{check_col_matching}{if the column names have information about tissue/
cell type and one-to-one correspondence of tissue types across species has to
be checked, set this parameter to be TRUE. Default FALSE.}

\item{col_sep}{separator in column names to separate different fields.
Example for column names 'hsapiens_brain', 'hsapiens_heart' etc.,
the separator is underscore.
Set it to NULL if column matching across species has to be
performed and there is no separator in the column names.
Only checked if check_col_matching = TRUE. Default underscore.}

\item{col_index}{If a separator separates information in column names,
the col_index is the index in the column name corresponding to tissue or
cell type. E.g. for column name 'hsapiens_brain', col_index is 2.
Only checked if check_col_matching = TRUE. Default NULL.}

\item{weighted}{If TRUE each Di^TDi is scaled using inverse variance weights
Default FALSE.}

\item{approximate}{TRUE will compute A-SBF. Default FALSE.}

\item{transform_matrix}{if TRUE, then Di will be transformed to compute
correlation matrix, and V is computed based on this instead of
Di^TDi. An unbiased estimate of covariance (denominator n-1) is
used for the computing correlation. Default FALSE.}

\item{verbose}{if TRUE print verbose lines. Default FALSE.}
}
\value{
a list containing u, delta, v, m, lambda (eigenvalues of m), and
other outputs of SBF/A-SBF factorization.
}
\description{
Function to compute Shared Basis Factorization (SBF) and
Approximate Shared Basis Factorization (A-SBF)
}
\examples{
# SBF call
avg_counts <- SBF::TissueExprSpecies
sbf <- SBF(matrix_list = avg_counts, check_col_matching = TRUE,
           col_index = 2, weighted = FALSE, approximate = FALSE,
           transform_matrix = FALSE)

# SBF call. Estimate V using inverse-variance weighted Di^TDi
sbf <- SBF(matrix_list = avg_counts, col_index = 2, weighted = TRUE,
           approximate = FALSE, transform_matrix = FALSE)

# SBF call using correlation matrix
avg_counts <- SBF::TissueExprSpecies
sbf_cor <- SBF(matrix_list = avg_counts, check_col_matching = TRUE,
           col_index = 2, weighted = FALSE, approximate = FALSE,
           transform_matrix = TRUE)

# A-SBF call
avg_counts <- SBF::TissueExprSpecies
asbf <- SBF(matrix_list = avg_counts, check_col_matching = TRUE,
            col_index = 2, weighted = FALSE,
            approximate = TRUE, transform_matrix = FALSE)
# calculate decomposition error
decomperror <- calcDecompError(avg_counts, asbf$u_ortho, asbf$delta, asbf$v)


# A-SBF call with inverse variance weighting
avg_counts <- SBF::TissueExprSpecies
asbf <- SBF(matrix_list = avg_counts, check_col_matching = TRUE,
            col_index = 2, weighted = TRUE,
            approximate = TRUE, transform_matrix = FALSE)
# calculate decomposition error
decomperror <- calcDecompError(avg_counts, asbf$u_ortho, asbf$delta, asbf$v)

# A-SBF call using correlation matrix
avg_counts <- SBF::TissueExprSpecies
asbf_cor <- SBF(matrix_list = avg_counts, check_col_matching = TRUE,
                col_index = 2, weighted = FALSE,
                approximate = TRUE, transform_matrix = TRUE)
# calculate decomposition error
decomperror <- calcDecompError(avg_counts, asbf_cor$u_ortho, asbf_cor$delta,
                                asbf_cor$v)
}
